% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_dgd.R
\name{schematic_g}
\alias{schematic_g}
\alias{schematic_Q}
\alias{gen_schematic}
\alias{make_schematic_grid}
\alias{plot_schematic}
\title{Schematic DGP
TODO: think about how to specify these, where they should live in the tlverse}
\usage{
schematic_g(W)

schematic_Q(A, W)

gen_schematic(n, eps_sd = 0.2)

make_schematic_grid()

plot_schematic(schematic, schematic_grid = NULL, schematic_meta = NULL)
}
\description{
Schematic DGP
TODO: think about how to specify these, where they should live in the tlverse
}
